+++ test/unit/context-construction-test.js
@@ -2,17 +2,6 @@ const zmq = require("../..")
 const semver = require("semver")
 const {assert} = require("chai")
 
-before(function() {
-  /* Avoid blocking on unclosed sockets on exit. */
-  if (semver.satisfies(zmq.version, ">= 4.2.0")) {
-    zmq.global.blocky = false
-  }
-})
-
-after(function() {
-  zmq.global.close()
-})
-
 describe("context construction", function() {
   afterEach(function() {
     gc()